Exposed function to set digital credentials presentment mechanism from intent (#1468)
